Refactor: #197: 임시 서비스 구현체 분리작업
